# Performance Tweaks to ContributeElementsToSparseMatrix