Several tweaks, but the end result is that this algorithm is now just as fast as before instead of a 133% slowdown on CPU.  

Added a check to not use atomics when in a serial build.  Improves speed by 3s out of 45s on benchmark

Assume components are in sorted order.  Improves speed by an additional 1s

Move col_idx population to its own loop.  Another 1s or so improvement.

Uses local scratch memory instead of heap memory pool.  No performance change, but removes need to set up memory pool

Fixed some compiler warnings when building for CUDA